[BE] FEAT: CQRS 패턴 적용하여 서비스 분리 #260
